+++ trunk/ecj/ec/app/push/Regression.java
@@ -164,7 +164,7 @@ public class Regression extends PushProblem implements SimpleProblemForm
                 
             // the fitness better be KozaFitness!
             KozaFitness f = ((KozaFitness)ind.fitness);
-            f.setStandardizedFitness(state,(float)sum);
+            f.setStandardizedFitness(state, sum);
             f.hits = hits;
             ind.evaluated = true;
             }
