+++ trunk/ecj/ec/app/tutorial2/AddSubtract.java
@@ -38,8 +38,7 @@ public class AddSubtract extends Problem implements SimpleProblemForm
             state.output.fatal("Whoa!  It's not a SimpleFitness!!!",null);
         ((SimpleFitness)ind2.fitness).setFitness(state,
             // what the heck, lets normalize the fitness for genome length
-            // so it's within float range
-            (float)(((double)rawfitness)/ind2.genome.length),
+            rawfitness/(double)ind2.genome.length,
             ///... is the individual ideal?  Indicate here...
             false);
         ind2.evaluated = true;
