+++ trunk/ecj/ec/app/tutorial4/MultiValuedRegression.java
@@ -58,7 +58,7 @@ public class MultiValuedRegression extends GPProblem implements SimpleProblemFor
 
             // the fitness better be KozaFitness!
             KozaFitness f = ((KozaFitness)ind.fitness);
-            f.setStandardizedFitness(state,(float)sum);
+            f.setStandardizedFitness(state, sum);
             f.hits = hits;
             ind.evaluated = true;
             }
